+++ composition/src/v1/normalization/types/types.ts
@@ -36,6 +36,7 @@ import {
   type INCLUDE_HEADERS,
   type MAX_AGE,
   type NEGATIVE_CACHE_TTL,
+  type OPENFED_CACHE_POPULATE,
   type OPENFED_ENTITY_CACHE,
   type PARTIAL_CACHE_LOAD,
   type SHADOW_MODE,
@@ -200,6 +201,20 @@ export type ShadowModeArgumentNode = {
   readonly kind: Kind.ARGUMENT;
   readonly name: NameNode & { readonly value: typeof SHADOW_MODE };
   readonly value: BooleanValueNode;
+  readonly loc?: Location;
+};
+
+export type CachePopulateDirectiveNode = {
+  readonly arguments: ReadonlyArray<CachePopulateArgumentNode>;
+  readonly kind: Kind.DIRECTIVE;
+  readonly name: NameNode & { readonly value: typeof OPENFED_CACHE_POPULATE };
+  readonly loc?: Location;
+};
+
+export type CachePopulateArgumentNode = {
+  readonly kind: Kind.ARGUMENT;
+  readonly name: NameNode & { readonly value: typeof MAX_AGE };
+  readonly value: IntValueNode;
   readonly loc?: Location;
 };